REFACTOR: int -> Long

+++ src/main/java/com/project/bumawiki/domain/docs/presentation/DocsInformationController.java
@@ -52,7 +52,7 @@ public class DocsInformationController {
     }
 
     @GetMapping("/find/version/{title}/different/{version}")
-    public VersionDocsDiffResponseDto showVersionDocsDiff(@PathVariable String title, @PathVariable int version) {
+    public VersionDocsDiffResponseDto showVersionDocsDiff(@PathVariable String title, @PathVariable Long version) {
         return docsInformationService.showVersionDocsDiff(title,version);
     }
